chore: break into reusable parts
