+++ packages/sanity-plugin-nacelle-input/.eslintrc.js
@@ -38,6 +38,7 @@ module.exports = {
       {
         aspects: ['invalidHref']
       }
-    ]
+    ],
+    'import/no-unresolved': [2, { ignore: ['^(all|part):'] }]
   }
 }
