+++ packages/sanity-plugin-nacelle-input/package.json
@@ -34,6 +34,7 @@
     "url": "https://github.com/getnacelle/nacellle-js/issues"
   },
   "dependencies": {
+    "@sanity/base": "^2.30.5",
     "isomorphic-unfetch": "^3.1.0",
     "prop-types": "^15.7.2",
     "swr": "^0.4.0"
@@ -54,7 +55,7 @@
     "styled-components": "^5.2.1"
   },
   "peerDependencies": {
-    "@sanity/base": "^2.0.9",
+    "@sanity/base": "^2.30.5",
     "@sanity/components": "^2.0.9",
     "@sanity/core": "^2.0.9",
     "@sanity/react-hooks": "^2.0.9",
