+++ packages/sanity-plugin-nacelle-input/sanity.json
@@ -5,13 +5,22 @@
   },
   "parts": [
     {
-      "name": "part:sanity-plugin-nacelle-input/nacelleInput",
+      "name": "part:sanity-plugin-nacelle-input/nacelleData",
       "path": "components/NacelleLinker.js",
       "description": "React component that provides product and/or collection data from Nacelle"
     },
     {
       "implements": "part:@sanity/base/schema-type",
-      "path": "schema.js"
+      "path": "nacelleData.js"
+    },
+    {
+      "name": "part:sanity-plugin-nacelle-input/nacelleReference",
+      "path": "components/NacelleLinker.js",
+      "description": "React component that provides product and/or collection data from Nacelle"
+    },
+    {
+      "implements": "part:@sanity/base/schema-type",
+      "path": "nacelleReference.js"
     }
   ]
 }
