+++ packages/sanity-plugin-nacelle-input/src/components/shared/Gallery.js
