+++ packages/sanity-plugin-nacelle-input/src/components/shared/Loading.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 
-import { useInterval } from '../hooks'
+import { useInterval } from '../../hooks'
 
 const Loading = () => {
   const [ellipses, setEllipses] = useState('.')
