+++ packages/sanity-plugin-nacelle-input/src/components/shared/NacelleDataFetcher.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 
-import { useNacelleStorefront } from '../hooks'
+import { useNacelleStorefront } from '../../hooks'
 import Gallery from './Gallery'
 import Loading from './Loading'
