+++ packages/sanity-plugin-nacelle-input/src/context/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-const HandleContext = React.createContext('')
+const ItemContext = React.createContext('')
 const SearchQueryContext = React.createContext('')
 const SpaceOptionsContext = React.createContext(null)
 
-export { HandleContext, SearchQueryContext, SpaceOptionsContext }
+export { ItemContext, SearchQueryContext, SpaceOptionsContext }
