+++ packages/sanity-plugin-nacelle-input/src/nacelleReference.js
@@ -1,0 +1,8 @@
+import NacelleLinker from './components/NacelleLinker'
+
+export default {
+  title: 'Nacelle Reference',
+  name: 'nacelleReference',
+  type: 'string',
+  inputComponent: NacelleLinker
+}
